moved getSuite
