+++ packages/vest/src/core/state/index.js
@@ -1,42 +1,10 @@
+import createStore from '../../lib/createStorage';
 import { KEY_CANCELED, KEY_SUITES } from './constants';
 
-export const { get, set, register } = (() => {
-  const storage = {
-    state: {
-      [KEY_SUITES]: {},
-      [KEY_CANCELED]: {},
-    },
-  };
-
-  /**
-   * Retrieves the state object or a portion of it.
-   */
-  const get = () => storage.state;
-
-  /**
-   * Updates the state with the value return from the setter callback.
-   * @param {Function} setter setter function.
-   * @returns {Object} updated state.
-   */
-  const set = setter => {
-    storage.state = setter(get());
-
-    return get();
-  };
-
-  const register = () =>
-    set(
-      state =>
-        state ?? {
-          [KEY_SUITES]: {},
-          [KEY_CANCELED]: {},
-        }
-    );
-
-  register();
-
-  return { get, set, register };
-})();
+export const { get, set, register } = createStore(() => ({
+  [KEY_SUITES]: {},
+  [KEY_CANCELED]: {},
+}));
 
 /**
  * Updates the state with the output of the setter callback.
@@ -50,9 +18,3 @@ export const setSuites = setter => {
   });
   return get()[KEY_SUITES];
 };
-
-/**
- * @param {string} suiteId.
- * @returns {Object} Suite from state.
- */
-export const getSuite = suiteId => get()[KEY_SUITES][suiteId];
